Switch to logging library, handle upload-aborting
Stdout logging was ok but if the user wanted to cancel the upload
process, they wouldn't be able to see when to do so since all console
prints were being piped to a file. Logging is now done with logging
library which can print to a file, but also to the console. More
user-friendly this way.

User now able to abort process mid-upload by pressing Ctrl-c.
The instructions say to 'hold' because pressing once catches the error
in gmusicapi, and only by pressing it twice does music_manager.py catch
it.