Use specific starting times, not just dates
The start_date config variable was just the date, not the time. This
meant multiple uploads in the same day would use midnight of the
start_date as the starting point for file checking, meaning previous
uploads from that day would be checked again. Not a huge deal since
Google can sniff out dupes, but it's now faster pre-upload. Note that
verbose mode date-setting will still use just dates, not times, so the
midnight rule will still apply. The fix in this commit applies to
subsequent, non-verbose uploads.

Also fixed log file names. There was colon (:) for the 24hr time format
which is invalid for Windows file names. Just removed the colon.